+++ crates/maple_core/src/searcher/grep/stoppable_searcher.rs
@@ -12,9 +12,9 @@ use std::time::{Duration, Instant};
 use tokio::sync::mpsc::{unbounded_channel, UnboundedSender};
 use types::{ProgressUpdate, Rank};
 
-const UPDATE_INTERVAL: Duration = Duration::from_millis(200);
+pub(super) const UPDATE_INTERVAL: Duration = Duration::from_millis(200);
 
-type SearcherMessage = crate::searcher::SearcherMessage<FileResult>;
+pub(super) type SearcherMessage = crate::searcher::SearcherMessage<FileResult>;
 
 #[derive(Debug, Default)]
 struct MatchEverything;
@@ -49,7 +49,7 @@ pub struct FileResult {
 }
 
 #[derive(Debug)]
-struct StoppableSearchImpl {
+pub(super) struct StoppableSearchImpl {
     paths: Vec<PathBuf>,
     matcher: Matcher,
     sender: UnboundedSender<SearcherMessage>,
@@ -57,7 +57,7 @@ struct StoppableSearchImpl {
 }
 
 impl StoppableSearchImpl {
-    fn new(
+    pub(super) fn new(
         paths: Vec<PathBuf>,
         matcher: Matcher,
         sender: UnboundedSender<SearcherMessage>,
@@ -71,7 +71,7 @@ impl StoppableSearchImpl {
         }
     }
 
-    fn run(self) {
+    pub(super) fn run(self) {
         let Self {
             paths,
             matcher,
@@ -161,57 +161,6 @@ impl StoppableSearchImpl {
                 WalkState::Continue
             })
         });
-    }
-}
-
-#[derive(Debug)]
-pub struct SearchResult {
-    pub matches: Vec<FileResult>,
-    pub total_matched: u64,
-    pub total_processed: u64,
-}
-
-pub async fn cli_search(paths: Vec<PathBuf>, matcher: Matcher) -> SearchResult {
-    let (sender, mut receiver) = unbounded_channel();
-
-    let stop_signal = Arc::new(AtomicBool::new(false));
-
-    std::thread::Builder::new()
-        .name("searcher-worker".into())
-        .spawn(move || StoppableSearchImpl::new(paths, matcher, sender, stop_signal).run())
-        .expect("Failed to spawn searcher worker thread");
-
-    let mut matches = Vec::new();
-    let mut total_matched = 0;
-    let mut total_processed = 0;
-
-    let mut past = Instant::now();
-
-    while let Some(searcher_message) = receiver.recv().await {
-        match searcher_message {
-            SearcherMessage::Match(file_result) => {
-                matches.push(file_result);
-                total_matched += 1;
-                total_processed += 1;
-            }
-            SearcherMessage::ProcessedOne => {
-                total_processed += 1;
-            }
-        }
-
-        if total_matched % 16 == 0 || total_processed % 16 == 0 {
-            let now = Instant::now();
-            if now > past + UPDATE_INTERVAL {
-                println!("total_matched: {total_matched:?}, total_processed: {total_processed:?}");
-                past = now;
-            }
-        }
-    }
-
-    SearchResult {
-        matches,
-        total_matched,
-        total_processed,
     }
 }
